wip(sdk): per-domain WIT scaffolding for migration
Checkpoint commit. Set up the foundation so a follow-up pass can finish
the wrapper migration:

contracts/ submodule
- Bumped from 30c6720 (users@1.0.0) to 324d4ab (per-domain + astrid:io).

astrid-sys
- Replaced single astrid-capsule.wit + monolithic 'capsule' world with
  a synthetic inline world that imports every per-domain host package
  and includes all four guest export worlds (interceptor, background,
  installable, upgradable).
- Added build.rs that stages contracts/host/*.wit into
  wit-staging/deps/astrid-<pkg>/ for wit-bindgen.
- Dropped serde derives from additional_derives — generated Resource
  types own kernel-side handles via Drop and can't be round-tripped
  through serde.
- Compiles clean for wasm32-wasip2.

astrid-sdk
- Includes the file split from local commit 5090ebf (lib.rs broken
  into elicit/http/identity/interceptors/ipc/kv/process per-file
  modules; request_response helper added to ipc).
- lib.rs imports rewritten for the per-domain layout:
  astrid_sys::astrid::<domain>::host as wit_<domain>, plus
  astrid_sys::astrid::io::{error, poll, streams} for the foundation
  primitives.

NOT yet done (handed off to follow-up pass):
- ~124 compile errors across the wrappers:
  - map_err(SysError::HostError) — host fns now return per-domain
    ErrorCode enums, not String. Needs a host_err helper or
    From<ErrorCode> impls.
  - wit_types::* references — types moved into per-domain modules.
  - Resource-backed handles (Subscription, FileHandle, TcpStream,
    ProcessHandle, HttpStream) — should become Drop-managed typed
    handles replacing the u64 ID pattern.
  - Record field renames (FileStat.mtime -> modified, is_dir -> kind
    on the new FileKind enum, IdentityCreateUserResponse shape).
- Macros (astrid-sdk-macros) — Guest trait now combines four worlds.
- examples/test-capsule update against new bindings.
- CHANGELOG entry.

Refs: core PR #752 (kernel-side per-domain migration).
