+++ astrid-sdk/src/lib.rs
@@ -40,11 +40,25 @@
 #![deny(clippy::unwrap_used)]
 #![cfg_attr(test, allow(clippy::unwrap_used))]
 
-use astrid_sys::astrid::capsule::{
-    approval as wit_approval, elicit as wit_elicit, fs as wit_fs, http as wit_http,
-    identity as wit_identity, ipc as wit_ipc, kv as wit_kv, net as wit_net, process as wit_process,
-    sys as wit_sys, types as wit_types, uplink as wit_uplink,
-};
+// Per-domain WIT (post-PR #752 split). Every host fn lives under
+// `astrid::<domain>::host`; the foundation I/O types are under
+// `astrid::io::{error, poll, streams}`. The aliases below preserve
+// the `wit_<domain>` names the wrappers use so most call sites stay
+// touched only in their error-mapping code.
+use astrid_sys::astrid::approval::host as wit_approval;
+use astrid_sys::astrid::elicit::host as wit_elicit;
+use astrid_sys::astrid::fs::host as wit_fs;
+use astrid_sys::astrid::http::host as wit_http;
+use astrid_sys::astrid::identity::host as wit_identity;
+use astrid_sys::astrid::io::error as wit_io_error;
+use astrid_sys::astrid::io::poll as wit_io_poll;
+use astrid_sys::astrid::io::streams as wit_io_streams;
+use astrid_sys::astrid::ipc::host as wit_ipc;
+use astrid_sys::astrid::kv::host as wit_kv;
+use astrid_sys::astrid::net::host as wit_net;
+use astrid_sys::astrid::process::host as wit_process;
+use astrid_sys::astrid::sys::host as wit_sys;
+use astrid_sys::astrid::uplink::host as wit_uplink;
 use borsh::{BorshDeserialize, BorshSerialize};
 use serde::{Deserialize, Serialize, de::DeserializeOwned};
 use thiserror::Error;
@@ -130,122 +144,7 @@ pub mod contracts {
 }
 
 /// Event bus messaging (like `std::sync::mpsc` but topic-based).
-pub mod ipc {
-    use super::*;
-
-    /// An active subscription to an IPC topic. Returned by [`subscribe`].
-    ///
-    /// Follows the typed-handle pattern used by [`crate::net::ListenerHandle`].
-    #[derive(Debug, Clone)]
-    pub struct SubscriptionHandle(pub(crate) u64);
-
-    impl SubscriptionHandle {
-        /// Raw handle ID for interop with lower-level APIs.
-        #[must_use]
-        pub fn id(&self) -> u64 {
-            self.0
-        }
-    }
-
-    /// Publish a UTF-8 string payload to an IPC topic.
-    ///
-    /// The IPC bus carries UTF-8 strings (WIT `string` type). For structured
-    /// data, use [`publish_json`] which handles serialization.
-    pub fn publish(topic: &str, payload: &str) -> Result<(), SysError> {
-        wit_ipc::ipc_publish(topic, payload).map_err(SysError::HostError)
-    }
-
-    /// Publish a JSON-serialized value to an IPC topic.
-    pub fn publish_json<T: Serialize>(topic: &str, payload: &T) -> Result<(), SysError> {
-        let json = serde_json::to_string(payload)?;
-        publish(topic, &json)
-    }
-
-    /// Publish a UTF-8 string payload on behalf of a specific principal.
-    ///
-    /// Like [`publish`] but stamps the outgoing envelope with the supplied
-    /// `principal` instead of this capsule's. Reserved for uplinks
-    /// (`uplink = true` in `Capsule.toml [capabilities]`) — the host
-    /// returns an error for any other caller. The trust model is "trust
-    /// the uplink": the kernel does not verify that the uplink actually
-    /// authenticated the claimed principal. See issue #658.
-    pub fn publish_as(topic: &str, payload: &str, principal: &str) -> Result<(), SysError> {
-        wit_ipc::ipc_publish_as(topic, payload, principal).map_err(SysError::HostError)
-    }
-
-    /// Publish a JSON-serialized value on behalf of a specific principal.
-    /// See [`publish_as`].
-    pub fn publish_json_as<T: Serialize>(
-        topic: &str,
-        payload: &T,
-        principal: &str,
-    ) -> Result<(), SysError> {
-        let json = serde_json::to_string(payload)?;
-        publish_as(topic, &json, principal)
-    }
-
-    /// Subscribe to an IPC topic. Returns a typed handle for polling/receiving.
-    pub fn subscribe(topic: &str) -> Result<SubscriptionHandle, SysError> {
-        let handle_id = wit_ipc::ipc_subscribe(topic).map_err(SysError::HostError)?;
-        Ok(SubscriptionHandle(handle_id))
-    }
-
-    pub fn unsubscribe(handle: &SubscriptionHandle) -> Result<(), SysError> {
-        wit_ipc::ipc_unsubscribe(handle.0).map_err(SysError::HostError)
-    }
-
-    /// A message received from the IPC bus.
-    #[derive(Debug, Clone)]
-    pub struct Message {
-        /// The topic this message was published on.
-        pub topic: String,
-        /// The message payload (JSON string).
-        pub payload: String,
-        /// UUID of the capsule that published this message.
-        pub source_id: String,
-    }
-
-    /// Result of polling or receiving from an IPC subscription.
-    #[derive(Debug)]
-    pub struct PollResult {
-        /// Messages received since the last poll/recv.
-        pub messages: Vec<Message>,
-        /// Number of messages dropped due to buffer overflow.
-        pub dropped: u64,
-        /// Cumulative lag (messages missed due to slow consumption).
-        pub lagged: u64,
-    }
-
-    /// Non-blocking poll for messages on a subscription.
-    pub fn poll(handle: &SubscriptionHandle) -> Result<PollResult, SysError> {
-        let envelope = wit_ipc::ipc_poll(handle.0).map_err(SysError::HostError)?;
-        Ok(envelope_to_poll_result(envelope))
-    }
-
-    /// Block until a message arrives on a subscription, or timeout.
-    ///
-    /// Max timeout is capped at 60,000 ms by the host.
-    pub fn recv(handle: &SubscriptionHandle, timeout_ms: u64) -> Result<PollResult, SysError> {
-        let envelope = wit_ipc::ipc_recv(handle.0, timeout_ms).map_err(SysError::HostError)?;
-        Ok(envelope_to_poll_result(envelope))
-    }
-
-    fn envelope_to_poll_result(envelope: wit_types::IpcEnvelope) -> PollResult {
-        PollResult {
-            messages: envelope
-                .messages
-                .into_iter()
-                .map(|m| Message {
-                    topic: m.topic,
-                    payload: m.payload,
-                    source_id: m.source_id,
-                })
-                .collect(),
-            dropped: envelope.dropped,
-            lagged: envelope.lagged,
-        }
-    }
-}
+pub mod ipc;
 
 /// Direct frontend messaging (uplinks to CLI, Telegram, etc.).
 pub mod uplink {
@@ -290,598 +189,11 @@ pub mod uplink {
 }
 
 /// The KV Airlock — Persistent Key-Value Storage
-pub mod kv {
-    use super::*;
-
-    pub fn get_bytes(key: &str) -> Result<Vec<u8>, SysError> {
-        let key_str = key;
-        let result = wit_kv::kv_get(key_str).map_err(SysError::HostError)?;
-        Ok(result.unwrap_or_default())
-    }
-
-    pub fn set_bytes(key: &str, value: &[u8]) -> Result<(), SysError> {
-        let key_str = key;
-        wit_kv::kv_set(key_str, value).map_err(SysError::HostError)
-    }
-
-    pub fn get_json<T: DeserializeOwned>(key: &str) -> Result<T, SysError> {
-        let bytes = get_bytes(key)?;
-        let parsed = serde_json::from_slice(&bytes)?;
-        Ok(parsed)
-    }
-
-    pub fn set_json<T: Serialize>(key: &str, value: &T) -> Result<(), SysError> {
-        let bytes = serde_json::to_vec(value)?;
-        set_bytes(key, &bytes)
-    }
-
-    /// Delete a key from the KV store.
-    ///
-    /// This is idempotent: deleting a non-existent key succeeds silently.
-    /// The underlying store returns whether the key existed, but that
-    /// information is not surfaced through the WASM host boundary.
-    pub fn delete(key: &str) -> Result<(), SysError> {
-        let key_str = key;
-        wit_kv::kv_delete(key_str).map_err(SysError::HostError)
-    }
-
-    /// List all keys matching a prefix.
-    ///
-    /// Returns an empty vec if no keys match. The prefix is matched
-    /// against key names within the capsule's scoped namespace.
-    pub fn list_keys(prefix: &str) -> Result<Vec<String>, SysError> {
-        let prefix_str = prefix;
-        wit_kv::kv_list_keys(prefix_str).map_err(SysError::HostError)
-    }
-
-    /// Delete all keys matching a prefix.
-    ///
-    /// Returns the number of keys deleted. The prefix is matched
-    /// against key names within the capsule's scoped namespace.
-    pub fn clear_prefix(prefix: &str) -> Result<u64, SysError> {
-        let prefix_str = prefix;
-        wit_kv::kv_clear_prefix(prefix_str).map_err(SysError::HostError)
-    }
-
-    pub fn get_borsh<T: BorshDeserialize>(key: &str) -> Result<T, SysError> {
-        let bytes = get_bytes(key)?;
-        let parsed = borsh::from_slice(&bytes)?;
-        Ok(parsed)
-    }
-
-    pub fn set_borsh<T: BorshSerialize>(key: &str, value: &T) -> Result<(), SysError> {
-        let bytes = borsh::to_vec(value)?;
-        set_bytes(key, &bytes)
-    }
-
-    // ---- Versioned KV helpers ----
-
-    /// Internal envelope for versioned KV data.
-    ///
-    /// Wire format: `{"__sv": <version>, "data": <payload>}`.
-    /// The `__sv` prefix is deliberately ugly to avoid collision with
-    /// user struct fields.
-    #[derive(Serialize, Deserialize)]
-    struct VersionedEnvelope<T> {
-        #[serde(rename = "__sv")]
-        schema_version: u32,
-        data: T,
-    }
-
-    /// Result of reading versioned data from KV.
-    #[derive(Debug)]
-    pub enum Versioned<T> {
-        /// Data is at the expected schema version.
-        Current(T),
-        /// Data is at an older version and needs migration.
-        NeedsMigration {
-            /// Raw JSON value of the `data` field.
-            raw: serde_json::Value,
-            /// The schema version that was stored.
-            stored_version: u32,
-        },
-        /// Key exists but data has no version envelope (pre-versioning legacy data).
-        Unversioned(serde_json::Value),
-        /// Key does not exist in KV.
-        NotFound,
-    }
-
-    /// Write versioned data to KV, wrapped in a schema-version envelope.
-    ///
-    /// The stored JSON looks like `{"__sv": 1, "data": { ... }}`.
-    /// Use [`get_versioned`] or [`get_versioned_or_migrate`] to read it back.
-    pub fn set_versioned<T: Serialize>(key: &str, value: &T, version: u32) -> Result<(), SysError> {
-        let envelope = VersionedEnvelope {
-            schema_version: version,
-            data: value,
-        };
-        set_json(key, &envelope)
-    }
-
-    /// Read versioned data from KV.
-    ///
-    /// Returns [`Versioned::Current`] if the stored version matches
-    /// `current_version`. Returns [`Versioned::NeedsMigration`] for older
-    /// versions. Returns an error for versions newer than `current_version`
-    /// (fail secure - don't silently interpret data from a schema you don't
-    /// understand).
-    ///
-    /// Data written by plain [`set_json`] (no envelope) returns
-    /// [`Versioned::Unversioned`].
-    pub fn get_versioned<T: DeserializeOwned>(
-        key: &str,
-        current_version: u32,
-    ) -> Result<Versioned<T>, SysError> {
-        let bytes = get_bytes(key)?;
-        parse_versioned(&bytes, current_version)
-    }
-
-    /// Core parsing logic for versioned KV data, separated from FFI for
-    /// testability. Operates on raw bytes as returned by `get_bytes`.
-    fn parse_versioned<T: DeserializeOwned>(
-        bytes: &[u8],
-        current_version: u32,
-    ) -> Result<Versioned<T>, SysError> {
-        // The host function `kv_get` returns an empty slice when the
-        // key is absent. A present key written via set_json/set_versioned
-        // always has at least the JSON envelope bytes, so empty = not found.
-        if bytes.is_empty() {
-            return Ok(Versioned::NotFound);
-        }
-
-        let mut value: serde_json::Value = serde_json::from_slice(bytes)?;
-
-        // Detect envelope by checking for __sv (u64) + data fields.
-        // If __sv is present but malformed (not a number, or missing data),
-        // return an error rather than silently treating as unversioned.
-        let sv_field = value.get("__sv");
-        let has_sv = sv_field.is_some();
-        let envelope_version = sv_field.and_then(|v| v.as_u64());
-        let has_data = value.get("data").is_some();
-
-        match (has_sv, envelope_version, has_data) {
-            // Valid envelope: __sv is a u64 and data is present.
-            // Take ownership of the data field via remove() to avoid cloning.
-            (_, Some(v), true) => {
-                let v = u32::try_from(v)
-                    .map_err(|_| SysError::ApiError("schema version exceeds u32::MAX".into()))?;
-                // Safety: the match guard confirmed has_data=true, so
-                // value is an object with a "data" key. This is infallible.
-                let data = value
-                    .as_object_mut()
-                    .and_then(|m| m.remove("data"))
-                    .expect("data field guaranteed by match condition");
-                if v == current_version {
-                    let parsed: T = serde_json::from_value(data)?;
-                    Ok(Versioned::Current(parsed))
-                } else if v < current_version {
-                    Ok(Versioned::NeedsMigration {
-                        raw: data,
-                        stored_version: v,
-                    })
-                } else {
-                    Err(SysError::ApiError(format!(
-                        "stored schema version {v} is newer than current \
-                         version {current_version} - cannot safely read"
-                    )))
-                }
-            }
-            // Malformed envelope: __sv present but data missing or __sv not a number.
-            (true, _, _) => Err(SysError::ApiError(
-                "malformed versioned envelope: __sv field present but \
-                 data field missing or __sv is not a number"
-                    .into(),
-            )),
-            // No __sv field at all: plain unversioned data.
-            (false, _, _) => Ok(Versioned::Unversioned(value)),
-        }
-    }
-
-    /// Read versioned data, automatically migrating older versions.
-    ///
-    /// `migrate_fn` receives the raw JSON and the stored version, and must
-    /// return a `T` at `current_version`. The migrated value is automatically
-    /// saved back to KV.
-    ///
-    /// **Warning:** The original data is overwritten after a successful
-    /// migration. If the write-back fails, the original data is preserved
-    /// and the migration will be re-attempted on the next call. Ensure
-    /// `migrate_fn` is idempotent and correct - there is no rollback
-    /// after a successful write.
-    ///
-    /// For [`Versioned::Unversioned`] data, `migrate_fn` is called with
-    /// version 0. For [`Versioned::NotFound`], returns `None`.
-    pub fn get_versioned_or_migrate<T: Serialize + DeserializeOwned>(
-        key: &str,
-        current_version: u32,
-        migrate_fn: impl FnOnce(serde_json::Value, u32) -> Result<T, SysError>,
-    ) -> Result<Option<T>, SysError> {
-        match get_versioned::<T>(key, current_version)? {
-            Versioned::Current(data) => Ok(Some(data)),
-            Versioned::NeedsMigration {
-                raw,
-                stored_version,
-            } => {
-                let migrated = migrate_fn(raw, stored_version)?;
-                set_versioned(key, &migrated, current_version)?;
-                Ok(Some(migrated))
-            }
-            Versioned::Unversioned(raw) => {
-                let migrated = migrate_fn(raw, 0)?;
-                set_versioned(key, &migrated, current_version)?;
-                Ok(Some(migrated))
-            }
-            Versioned::NotFound => Ok(None),
-        }
-    }
-
-    #[cfg(test)]
-    mod tests {
-        use super::*;
-
-        #[derive(Debug, Serialize, Deserialize, PartialEq)]
-        struct TestData {
-            name: String,
-            count: u32,
-        }
-
-        // ---- Envelope serialization tests ----
-
-        #[test]
-        fn versioned_envelope_roundtrip() {
-            let envelope = VersionedEnvelope {
-                schema_version: 1,
-                data: TestData {
-                    name: "hello".into(),
-                    count: 42,
-                },
-            };
-            let json = serde_json::to_string(&envelope).unwrap();
-            assert!(json.contains("\"__sv\":1"));
-            assert!(json.contains("\"data\":{"));
-
-            let parsed: VersionedEnvelope<TestData> = serde_json::from_str(&json).unwrap();
-            assert_eq!(parsed.schema_version, 1);
-            assert_eq!(
-                parsed.data,
-                TestData {
-                    name: "hello".into(),
-                    count: 42,
-                }
-            );
-        }
-
-        #[test]
-        fn versioned_envelope_wire_format() {
-            let envelope = VersionedEnvelope {
-                schema_version: 3,
-                data: serde_json::json!({"key": "value"}),
-            };
-            let json = serde_json::to_string(&envelope).unwrap();
-            let parsed: serde_json::Value = serde_json::from_str(&json).unwrap();
-
-            assert_eq!(parsed["__sv"], 3);
-            assert_eq!(parsed["data"]["key"], "value");
-        }
-
-        // ---- parse_versioned logic tests ----
-
-        #[test]
-        fn parse_versioned_empty_bytes_returns_not_found() {
-            let result = parse_versioned::<TestData>(b"", 1).unwrap();
-            assert!(matches!(result, Versioned::NotFound));
-        }
-
-        #[test]
-        fn parse_versioned_current_version_returns_current() {
-            let bytes = br#"{"__sv":2,"data":{"name":"hello","count":42}}"#;
-            let result = parse_versioned::<TestData>(bytes, 2).unwrap();
-            match result {
-                Versioned::Current(data) => {
-                    assert_eq!(data.name, "hello");
-                    assert_eq!(data.count, 42);
-                }
-                other => panic!("expected Current, got {other:?}"),
-            }
-        }
-
-        #[test]
-        fn parse_versioned_older_version_returns_needs_migration() {
-            let bytes = br#"{"__sv":1,"data":{"name":"old","count":1}}"#;
-            let result = parse_versioned::<TestData>(bytes, 3).unwrap();
-            match result {
-                Versioned::NeedsMigration {
-                    raw,
-                    stored_version,
-                } => {
-                    assert_eq!(stored_version, 1);
-                    assert_eq!(raw["name"], "old");
-                    assert_eq!(raw["count"], 1);
-                }
-                other => panic!("expected NeedsMigration, got {other:?}"),
-            }
-        }
-
-        #[test]
-        fn parse_versioned_newer_version_returns_error() {
-            let bytes = br#"{"__sv":5,"data":{"name":"future","count":0}}"#;
-            let result = parse_versioned::<TestData>(bytes, 2);
-            assert!(result.is_err());
-            let err = result.unwrap_err().to_string();
-            assert!(
-                err.contains("newer than current"),
-                "error should mention newer version: {err}"
-            );
-        }
-
-        #[test]
-        fn parse_versioned_plain_json_returns_unversioned() {
-            let bytes = br#"{"name":"legacy","count":99}"#;
-            let result = parse_versioned::<TestData>(bytes, 1).unwrap();
-            match result {
-                Versioned::Unversioned(val) => {
-                    assert_eq!(val["name"], "legacy");
-                    assert_eq!(val["count"], 99);
-                }
-                other => panic!("expected Unversioned, got {other:?}"),
-            }
-        }
-
-        #[test]
-        fn parse_versioned_malformed_sv_without_data_returns_error() {
-            let bytes = br#"{"__sv":1,"payload":"something"}"#;
-            let result = parse_versioned::<TestData>(bytes, 1);
-            assert!(result.is_err());
-            let err = result.unwrap_err().to_string();
-            assert!(
-                err.contains("malformed"),
-                "error should mention malformed envelope: {err}"
-            );
-        }
-
-        #[test]
-        fn parse_versioned_non_numeric_sv_returns_error() {
-            let bytes = br#"{"__sv":"one","data":{}}"#;
-            let result = parse_versioned::<TestData>(bytes, 1);
-            assert!(result.is_err());
-            let err = result.unwrap_err().to_string();
-            assert!(
-                err.contains("malformed"),
-                "error should mention malformed envelope: {err}"
-            );
-        }
-
-        #[test]
-        fn parse_versioned_version_zero_is_valid() {
-            // Version 0 is a legitimate version (initial schema).
-            let bytes = br#"{"__sv":0,"data":{"name":"v0","count":0}}"#;
-            let result = parse_versioned::<TestData>(bytes, 0).unwrap();
-            assert!(matches!(result, Versioned::Current(_)));
-        }
-
-        #[test]
-        fn parse_versioned_invalid_json_returns_error() {
-            let result = parse_versioned::<TestData>(b"not json", 1);
-            assert!(result.is_err());
-        }
-    }
-}
+pub mod kv;
 
 /// The HTTP Airlock — External Network Requests
 /// Outbound HTTP — typed request API over the host HTTP airlock.
-pub mod http {
-    use super::*;
-    use serde::Serialize;
-    use std::collections::HashMap;
-
-    /// An HTTP request.
-    ///
-    /// Construct via [`Request::get`], [`Request::post`], etc. or
-    /// [`Request::new`] for arbitrary methods.
-    #[derive(Debug, Clone, Serialize)]
-    pub struct Request {
-        url: String,
-        method: String,
-        #[serde(default, skip_serializing_if = "HashMap::is_empty")]
-        headers: HashMap<String, String>,
-        #[serde(default, skip_serializing_if = "Option::is_none")]
-        body: Option<String>,
-    }
-
-    impl Request {
-        /// Create a request with an arbitrary method.
-        pub fn new(method: impl Into<String>, url: impl Into<String>) -> Self {
-            Self {
-                url: url.into(),
-                method: method.into(),
-                headers: HashMap::new(),
-                body: None,
-            }
-        }
-
-        /// Create a GET request.
-        pub fn get(url: impl Into<String>) -> Self {
-            Self::new("GET", url)
-        }
-
-        /// Create a POST request.
-        pub fn post(url: impl Into<String>) -> Self {
-            Self::new("POST", url)
-        }
-
-        /// Create a PUT request.
-        pub fn put(url: impl Into<String>) -> Self {
-            Self::new("PUT", url)
-        }
-
-        /// Create a DELETE request.
-        pub fn delete(url: impl Into<String>) -> Self {
-            Self::new("DELETE", url)
-        }
-
-        /// Add a header.
-        pub fn header(mut self, key: impl Into<String>, value: impl Into<String>) -> Self {
-            self.headers.insert(key.into(), value.into());
-            self
-        }
-
-        /// Set the request body.
-        pub fn body(mut self, body: impl Into<String>) -> Self {
-            self.body = Some(body.into());
-            self
-        }
-
-        /// Set a JSON body (serializes the value and sets Content-Type).
-        pub fn json<T: Serialize>(self, value: &T) -> Result<Self, SysError> {
-            let json = serde_json::to_string(value)?;
-            Ok(self.header("Content-Type", "application/json").body(json))
-        }
-
-        /// Convert to the WIT HttpRequestData type.
-        fn to_wit(&self) -> wit_types::HttpRequestData {
-            wit_types::HttpRequestData {
-                url: self.url.clone(),
-                method: self.method.clone(),
-                headers: self
-                    .headers
-                    .iter()
-                    .map(|(k, v)| wit_types::KeyValuePair {
-                        key: k.clone(),
-                        value: v.clone(),
-                    })
-                    .collect(),
-                body: self.body.clone(),
-            }
-        }
-    }
-
-    /// An HTTP response from a non-streaming request.
-    ///
-    /// All fields are private — use accessor methods to read them.
-    #[derive(Debug)]
-    pub struct Response {
-        status: u16,
-        headers: HashMap<String, String>,
-        body: Vec<u8>,
-    }
-
-    impl Response {
-        /// HTTP status code (e.g. 200, 404, 500).
-        #[must_use]
-        pub fn status(&self) -> u16 {
-            self.status
-        }
-
-        /// Response headers.
-        #[must_use]
-        pub fn headers(&self) -> &HashMap<String, String> {
-            &self.headers
-        }
-
-        /// The raw response body as bytes.
-        #[must_use]
-        pub fn bytes(&self) -> &[u8] {
-            &self.body
-        }
-
-        /// The response body as a UTF-8 string.
-        pub fn text(&self) -> Result<&str, SysError> {
-            core::str::from_utf8(&self.body).map_err(|e| SysError::ApiError(e.to_string()))
-        }
-
-        /// Deserialize the response body as JSON.
-        pub fn json<T: serde::de::DeserializeOwned>(&self) -> Result<T, SysError> {
-            serde_json::from_slice(&self.body).map_err(SysError::from)
-        }
-
-        /// Whether the status code indicates success (2xx).
-        #[must_use]
-        pub fn is_success(&self) -> bool {
-            (200..300).contains(&self.status)
-        }
-    }
-
-    /// Send an HTTP request and wait for the full response.
-    pub fn send(request: &Request) -> Result<Response, SysError> {
-        let wit_req = request.to_wit();
-        let result = wit_http::http_request(&wit_req).map_err(SysError::HostError)?;
-        let headers: HashMap<String, String> = result
-            .headers
-            .into_iter()
-            .map(|kv| (kv.key, kv.value))
-            .collect();
-        Ok(Response {
-            status: result.status,
-            headers,
-            body: result.body,
-        })
-    }
-
-    /// Represents an active streaming HTTP response.
-    ///
-    /// Must be explicitly closed via [`stream_close`] when done.
-    /// Not `Clone` — each handle is a unique owner of the host-side resource.
-    #[derive(Debug)]
-    pub struct HttpStreamHandle(pub(crate) u64);
-
-    impl HttpStreamHandle {
-        /// Raw handle ID for interop with lower-level APIs.
-        #[must_use]
-        pub fn id(&self) -> u64 {
-            self.0
-        }
-    }
-
-    /// Metadata returned when a streaming HTTP request is initiated.
-    pub struct StreamStartResponse {
-        /// The handle to use for subsequent [`stream_read`] / [`stream_close`] calls.
-        pub handle: HttpStreamHandle,
-        /// HTTP status code.
-        pub status: u16,
-        /// Response headers.
-        pub headers: HashMap<String, String>,
-    }
-
-    /// Start a streaming HTTP request.
-    ///
-    /// Sends the request and waits for the status/headers to arrive.
-    /// Returns a [`StreamStartResponse`] with the handle, status, and headers.
-    /// Use [`stream_read`] to consume the body in chunks.
-    pub fn stream_start(request: &Request) -> Result<StreamStartResponse, SysError> {
-        let wit_req = request.to_wit();
-        let result = wit_http::http_stream_start(&wit_req).map_err(SysError::HostError)?;
-        let headers: HashMap<String, String> = result
-            .headers
-            .into_iter()
-            .map(|kv| (kv.key, kv.value))
-            .collect();
-        Ok(StreamStartResponse {
-            handle: HttpStreamHandle(result.handle),
-            status: result.status,
-            headers,
-        })
-    }
-
-    /// Read the next chunk from a streaming HTTP response.
-    ///
-    /// Returns `Ok(Some(bytes))` with the next chunk of data, or
-    /// `Ok(None)` when the stream is exhausted (EOF).
-    pub fn stream_read(stream: &HttpStreamHandle) -> Result<Option<Vec<u8>>, SysError> {
-        let result = wit_http::http_stream_read(stream.0).map_err(SysError::HostError)?;
-        if result.is_empty() {
-            Ok(None)
-        } else {
-            Ok(Some(result))
-        }
-    }
-
-    /// Close a streaming HTTP response, releasing host-side resources.
-    ///
-    /// Idempotent — closing an already-closed handle is a no-op.
-    pub fn stream_close(stream: &HttpStreamHandle) -> Result<(), SysError> {
-        wit_http::http_stream_close(stream.0).map_err(SysError::HostError)
-    }
-}
+pub mod http;
 
 /// Capsule configuration (like `std::env`).
 ///
@@ -1038,256 +350,14 @@ pub mod capabilities {
 }
 
 pub mod net;
-pub mod process {
-    use super::*;
-    use serde::Deserialize;
-
-    /// Result returned from a spawned host process.
-    #[derive(Debug, Deserialize)]
-    pub struct ProcessResult {
-        pub stdout: String,
-        pub stderr: String,
-        pub exit_code: i32,
-    }
-
-    /// Spawns a native host process (blocks until completion).
-    /// The Capsule must have the `host_process` capability granted for this command.
-    pub fn spawn(cmd: &str, args: &[&str]) -> Result<ProcessResult, SysError> {
-        let request = wit_types::SpawnRequest {
-            cmd: cmd.to_string(),
-            args: args.iter().map(|s| s.to_string()).collect(),
-        };
-        let result = wit_process::spawn(&request).map_err(SysError::HostError)?;
-        Ok(ProcessResult {
-            stdout: result.stdout,
-            stderr: result.stderr,
-            exit_code: result.exit_code,
-        })
-    }
-
-    // -------------------------------------------------------------------
-    // Background process management
-    // -------------------------------------------------------------------
-
-    /// Handle returned when a background process is spawned.
-    #[derive(Debug, Deserialize)]
-    pub struct BackgroundProcessHandle {
-        /// Opaque handle ID (not an OS PID).
-        pub(crate) id: u64,
-    }
-
-    impl BackgroundProcessHandle {
-        /// Returns the opaque handle ID for this process.
-        pub fn id(&self) -> u64 {
-            self.id
-        }
-    }
-
-    /// Buffered logs and status from a background process.
-    #[derive(Debug, Deserialize)]
-    pub struct ProcessLogs {
-        /// New stdout output since the last read.
-        pub stdout: String,
-        /// New stderr output since the last read.
-        pub stderr: String,
-        /// Whether the process is still running.
-        pub running: bool,
-        /// Exit code if the process has exited.
-        pub exit_code: Option<i32>,
-    }
-
-    /// Result from killing a background process.
-    #[derive(Debug, Deserialize)]
-    pub struct KillResult {
-        /// Whether the process was successfully killed.
-        pub killed: bool,
-        /// Exit code of the terminated process.
-        pub exit_code: Option<i32>,
-        /// Any remaining buffered stdout.
-        pub stdout: String,
-        /// Any remaining buffered stderr.
-        pub stderr: String,
-    }
-
-    /// Spawn a background host process.
-    ///
-    /// Returns an opaque handle that can be used with [`read_logs`] and
-    /// [`kill`]. The process runs sandboxed with piped stdout/stderr.
-    pub fn spawn_background(cmd: &str, args: &[&str]) -> Result<BackgroundProcessHandle, SysError> {
-        let request = wit_types::SpawnRequest {
-            cmd: cmd.to_string(),
-            args: args.iter().map(|s| s.to_string()).collect(),
-        };
-        let result = wit_process::spawn_background(&request).map_err(SysError::HostError)?;
-        Ok(BackgroundProcessHandle { id: result.id })
-    }
-
-    /// Read buffered output from a background process.
-    ///
-    /// Each call drains the buffer and returns only NEW output since the
-    /// last read. Also reports whether the process is still running.
-    pub fn read_logs(id: u64) -> Result<ProcessLogs, SysError> {
-        let result = wit_process::read_logs(id).map_err(SysError::HostError)?;
-        Ok(ProcessLogs {
-            stdout: result.stdout,
-            stderr: result.stderr,
-            running: result.running,
-            exit_code: result.exit_code,
-        })
-    }
-
-    /// Kill a background process and release its resources.
-    ///
-    /// Returns any remaining buffered output along with the exit code.
-    pub fn kill(id: u64) -> Result<KillResult, SysError> {
-        let result = wit_process::kill(id).map_err(SysError::HostError)?;
-        Ok(KillResult {
-            killed: result.killed,
-            exit_code: result.exit_code,
-            stdout: result.stdout,
-            stderr: result.stderr,
-        })
-    }
-}
+pub mod process;
 
 /// The Elicit Airlock - User Input During Install/Upgrade Lifecycle
 ///
 /// These functions are only callable during `#[astrid::install]` and
 /// `#[astrid::upgrade]` hooks. Calling them from a tool or interceptor
 /// returns a host error.
-pub mod elicit {
-    use super::*;
-
-    /// Validates that the elicit key is non-empty and not whitespace-only.
-    fn validate_key(key: &str) -> Result<(), SysError> {
-        if key.trim().is_empty() {
-            return Err(SysError::ApiError("elicit key must not be empty".into()));
-        }
-        Ok(())
-    }
-
-    /// Store a secret via the kernel's `SecretStore`. The capsule **never**
-    /// receives the value. Returns `Ok(())` confirming the user provided it.
-    pub fn secret(key: &str, description: &str) -> Result<(), SysError> {
-        validate_key(key)?;
-        let req = wit_types::ElicitRequest {
-            elicit_type: "secret".to_string(),
-            key: key.to_string(),
-            description: description.to_string(),
-            options: None,
-            default_value: None,
-        };
-        let resp_str = wit_elicit::elicit(&req).map_err(SysError::HostError)?;
-
-        #[derive(serde::Deserialize)]
-        struct SecretResp {
-            ok: bool,
-        }
-        let resp: SecretResp = serde_json::from_str(&resp_str)?;
-        if !resp.ok {
-            return Err(SysError::ApiError(
-                "kernel did not confirm secret storage".into(),
-            ));
-        }
-        Ok(())
-    }
-
-    /// Check if a secret has been configured (without reading it).
-    pub fn has_secret(key: &str) -> Result<bool, SysError> {
-        validate_key(key)?;
-        wit_elicit::has_secret(key).map_err(SysError::HostError)
-    }
-
-    /// Shared implementation for text elicitation with optional default.
-    fn elicit_text(
-        key: &str,
-        description: &str,
-        default: Option<&str>,
-    ) -> Result<String, SysError> {
-        validate_key(key)?;
-        let req = wit_types::ElicitRequest {
-            elicit_type: "text".to_string(),
-            key: key.to_string(),
-            description: description.to_string(),
-            options: None,
-            default_value: default.map(|s| s.to_string()),
-        };
-        let resp_str = wit_elicit::elicit(&req).map_err(SysError::HostError)?;
-
-        #[derive(serde::Deserialize)]
-        struct TextResp {
-            value: String,
-        }
-        let resp: TextResp = serde_json::from_str(&resp_str)?;
-        Ok(resp.value)
-    }
-
-    /// Prompt for a text value. Blocks until the user responds.
-    /// Use [`secret()`] for sensitive data - this returns the value to the capsule.
-    pub fn text(key: &str, description: &str) -> Result<String, SysError> {
-        elicit_text(key, description, None)
-    }
-
-    /// Prompt with a default value pre-filled.
-    pub fn text_with_default(
-        key: &str,
-        description: &str,
-        default: &str,
-    ) -> Result<String, SysError> {
-        elicit_text(key, description, Some(default))
-    }
-
-    /// Prompt for a selection from a list. Returns the selected value.
-    pub fn select(key: &str, description: &str, options: &[&str]) -> Result<String, SysError> {
-        validate_key(key)?;
-        if options.is_empty() {
-            return Err(SysError::ApiError(
-                "select requires at least one option".into(),
-            ));
-        }
-        let req = wit_types::ElicitRequest {
-            elicit_type: "select".to_string(),
-            key: key.to_string(),
-            description: description.to_string(),
-            options: Some(options.iter().map(|s| s.to_string()).collect()),
-            default_value: None,
-        };
-        let resp_str = wit_elicit::elicit(&req).map_err(SysError::HostError)?;
-
-        #[derive(serde::Deserialize)]
-        struct SelectResp {
-            value: String,
-        }
-        let resp: SelectResp = serde_json::from_str(&resp_str)?;
-        if !options.iter().any(|o| *o == resp.value) {
-            let truncated: String = resp.value.chars().take(64).collect();
-            return Err(SysError::ApiError(format!(
-                "host returned value '{truncated}' not in provided options",
-            )));
-        }
-        Ok(resp.value)
-    }
-
-    /// Prompt for multiple text values (array input).
-    pub fn array(key: &str, description: &str) -> Result<Vec<String>, SysError> {
-        validate_key(key)?;
-        let req = wit_types::ElicitRequest {
-            elicit_type: "array".to_string(),
-            key: key.to_string(),
-            description: description.to_string(),
-            options: None,
-            default_value: None,
-        };
-        let resp_str = wit_elicit::elicit(&req).map_err(SysError::HostError)?;
-
-        #[derive(serde::Deserialize)]
-        struct ArrayResp {
-            values: Vec<String>,
-        }
-        let resp: ArrayResp = serde_json::from_str(&resp_str)?;
-        Ok(resp.values)
-    }
-}
+pub mod elicit;
 
 /// Auto-subscribed interceptor bindings for run-loop capsules.
 ///
@@ -1295,69 +365,7 @@ pub mod elicit {
 /// auto-subscribes to each interceptor's topic and delivers events through
 /// the IPC channel the run loop already reads from. This module provides
 /// helpers to query the subscription mappings and dispatch events by action.
-pub mod interceptors {
-    use super::*;
-
-    /// A single interceptor subscription binding.
-    #[derive(Debug)]
-    pub struct InterceptorBinding {
-        /// The IPC subscription handle ID.
-        pub handle_id: u64,
-        /// The interceptor action name from the manifest.
-        pub action: String,
-        /// The event topic this interceptor subscribes to.
-        pub topic: String,
-    }
-
-    impl InterceptorBinding {
-        /// Return a subscription handle for use with [`ipc::poll`] / [`ipc::recv`].
-        #[must_use]
-        pub fn subscription_handle(&self) -> ipc::SubscriptionHandle {
-            ipc::SubscriptionHandle(self.handle_id)
-        }
-
-        /// Return the raw handle ID bytes (for lower-level interop).
-        #[must_use]
-        pub fn handle_bytes(&self) -> Vec<u8> {
-            self.handle_id.to_string().into_bytes()
-        }
-    }
-
-    /// Query the runtime for auto-subscribed interceptor handles.
-    ///
-    /// Returns an empty vec if this capsule has no auto-subscribed interceptors
-    /// (i.e. it does not have both `run()` and `[[interceptor]]`).
-    pub fn bindings() -> Result<Vec<InterceptorBinding>, SysError> {
-        let handles = wit_ipc::get_interceptor_handles().map_err(SysError::HostError)?;
-        Ok(handles
-            .into_iter()
-            .map(|h| InterceptorBinding {
-                handle_id: h.handle_id,
-                action: h.action,
-                topic: h.topic,
-            })
-            .collect())
-    }
-
-    /// Poll all interceptor subscriptions and dispatch pending events.
-    ///
-    /// For each binding with pending messages, calls
-    /// `handler(action, messages)` once with the typed [`ipc::PollResult`].
-    /// Bindings with no pending messages are skipped.
-    pub fn poll(
-        bindings: &[InterceptorBinding],
-        mut handler: impl FnMut(&str, &ipc::PollResult),
-    ) -> Result<(), SysError> {
-        for binding in bindings {
-            let handle = binding.subscription_handle();
-            let result = ipc::poll(&handle)?;
-            if !result.messages.is_empty() {
-                handler(&binding.action, &result);
-            }
-        }
-        Ok(())
-    }
-}
+pub mod interceptors;
 
 /// Request human approval for sensitive actions from within a capsule.
 ///
@@ -1385,166 +393,7 @@ pub mod interceptors {
 /// - `["resolve"]` - resolve platform users
 /// - `["link"]` - resolve, link, unlink, and list links
 /// - `["admin"]` - all of the above plus create new users
-pub mod identity {
-    use super::*;
-
-    /// A resolved Astrid user returned by [`resolve`].
-    #[derive(Debug)]
-    pub struct ResolvedUser {
-        /// The Astrid-native user ID (UUID).
-        pub user_id: String,
-        /// Optional display name.
-        pub display_name: Option<String>,
-    }
-
-    /// A platform-to-Astrid identity link.
-    #[derive(Debug)]
-    pub struct Link {
-        /// Platform name (e.g. "discord", "twitch").
-        pub platform: String,
-        /// Platform-specific user identifier.
-        pub platform_user_id: String,
-        /// The Astrid user this is linked to.
-        pub astrid_user_id: String,
-        /// When the link was created (RFC 3339).
-        pub linked_at: String,
-        /// How the link was established (e.g. "system", "chat_command").
-        pub method: String,
-    }
-
-    /// Resolve a platform user to an Astrid user.
-    ///
-    /// Returns `Ok(Some(user))` if the platform identity is linked,
-    /// `Ok(None)` if not found. Requires `identity = ["resolve"]` or higher.
-    pub fn resolve(
-        platform: &str,
-        platform_user_id: &str,
-    ) -> Result<Option<ResolvedUser>, SysError> {
-        let request = wit_types::IdentityResolveRequest {
-            platform: platform.to_string(),
-            platform_user_id: platform_user_id.to_string(),
-        };
-        let resp = wit_identity::identity_resolve(&request).map_err(SysError::HostError)?;
-        if resp.found {
-            let user_id = resp.user_id.ok_or_else(|| {
-                SysError::ApiError("host returned found=true but user_id was missing".into())
-            })?;
-            Ok(Some(ResolvedUser {
-                user_id,
-                display_name: resp.display_name,
-            }))
-        } else if let Some(err) = resp.error {
-            Err(SysError::ApiError(err))
-        } else {
-            Ok(None)
-        }
-    }
-
-    /// Link a platform identity to an Astrid user.
-    ///
-    /// - `method` describes how the link was established (e.g. "chat_command", "system").
-    ///
-    /// Requires `identity = ["link"]` or higher.
-    pub fn link(
-        platform: &str,
-        platform_user_id: &str,
-        astrid_user_id: &str,
-        method: &str,
-    ) -> Result<(), SysError> {
-        let request = wit_types::IdentityLinkRequest {
-            platform: platform.to_string(),
-            platform_user_id: platform_user_id.to_string(),
-            astrid_user_id: astrid_user_id.to_string(),
-            method: method.to_string(),
-        };
-        let resp = wit_identity::identity_link(&request).map_err(SysError::HostError)?;
-        if !resp.ok {
-            return Err(SysError::ApiError(
-                resp.error.unwrap_or_else(|| "identity link failed".into()),
-            ));
-        }
-        Ok(())
-    }
-
-    /// Unlink a platform identity from its Astrid user.
-    ///
-    /// Returns `true` if a link was removed, `false` if none existed.
-    /// Requires `identity = ["link"]` or higher.
-    pub fn unlink(platform: &str, platform_user_id: &str) -> Result<bool, SysError> {
-        let request = wit_types::IdentityUnlinkRequest {
-            platform: platform.to_string(),
-            platform_user_id: platform_user_id.to_string(),
-        };
-        let resp = wit_identity::identity_unlink(&request).map_err(SysError::HostError)?;
-        if !resp.ok {
-            return Err(SysError::ApiError(
-                resp.error
-                    .unwrap_or_else(|| "identity unlink failed".into()),
-            ));
-        }
-        Ok(resp.removed.unwrap_or(false))
-    }
-
-    /// Create a new Astrid user.
-    ///
-    /// Returns the UUID of the newly created user.
-    /// Requires `identity = ["admin"]`.
-    pub fn create_user(display_name: Option<&str>) -> Result<String, SysError> {
-        let request = wit_types::IdentityCreateUserRequest {
-            display_name: display_name.map(|s| s.to_string()),
-        };
-        let resp = wit_identity::identity_create_user(&request).map_err(SysError::HostError)?;
-        if !resp.ok {
-            return Err(SysError::ApiError(
-                resp.error
-                    .unwrap_or_else(|| "identity create_user failed".into()),
-            ));
-        }
-        resp.user_id
-            .ok_or_else(|| SysError::ApiError("missing user_id in response".into()))
-    }
-
-    /// List all platform links for an Astrid user.
-    ///
-    /// Returns all linked platform identities for the given user UUID.
-    /// Requires `identity = ["link"]` or higher.
-    pub fn list_links(astrid_user_id: &str) -> Result<Vec<Link>, SysError> {
-        let request = wit_types::IdentityListLinksRequest {
-            astrid_user_id: astrid_user_id.to_string(),
-        };
-        let resp = wit_identity::identity_list_links(&request).map_err(SysError::HostError)?;
-        if !resp.ok {
-            return Err(SysError::ApiError(
-                resp.error
-                    .unwrap_or_else(|| "identity list_links failed".into()),
-            ));
-        }
-        // Parse links from the links_json field.
-        if let Some(json_str) = &resp.links_json {
-            #[derive(Deserialize)]
-            struct LinkInfo {
-                platform: String,
-                platform_user_id: String,
-                astrid_user_id: String,
-                linked_at: String,
-                method: String,
-            }
-            let links: Vec<LinkInfo> = serde_json::from_str(json_str)?;
-            Ok(links
-                .into_iter()
-                .map(|l| Link {
-                    platform: l.platform,
-                    platform_user_id: l.platform_user_id,
-                    astrid_user_id: l.astrid_user_id,
-                    linked_at: l.linked_at,
-                    method: l.method,
-                })
-                .collect())
-        } else {
-            Ok(Vec::new())
-        }
-    }
-}
+pub mod identity;
 
 /// Human-in-the-loop approval for sensitive actions.
 ///
