+++ astrid-sys/.gitignore
@@ -1,0 +1,1 @@
+wit-staging/
